refactor: organized the cart event listener
